# Conflicts:
#	panda/board/safety/safety_gm.h
#	selfdrive/manager/process_config.py
#	selfdrive/sentry.py
#	selfdrive/ui/qt/onroad.cc

+++ panda/board/safety/safety_gm.h
@@ -334,7 +334,7 @@ static int gm_fwd_hook(int bus_num, CANPacket_t *to_fwd) {
 }
 
 
-static const addr_checks* gm_init(uint32_t param) {
+static const addr_checks* gm_init(uint16_t param) {
   UNUSED(param);
   controls_allowed = false;
   relay_malfunction_reset();
